@@ -78,46 +80,52 @@
     "compression": "1.7.4",
     "cookie-parser": "1.4.6",
     "ejs": "3.1.9",
-    "graphql": "16.6.0",
+    "graphql": "16.7.1",
     "graphql-query-complexity": "0.12.0",
     "graphql-subscriptions": "2.0.0",
     "graphql-upload": "15.0.2",
     "js-sha256": "0.9.0",
     "json-to-graphql-query": "2.2.5",
-    "light-my-request": "5.9.1",
+    "light-my-request": "5.10.0",
     "lodash": "4.17.21",
-    "mongodb": "4.14.0",
-    "mongoose": "6.10.4",
+    "mongodb": "4.16.0",
+    "mongoose": "6.11.3",
     "mongoose-gridfs": "1.3.0",
     "multer": "1.4.5-lts.1",
     "multer-gridfs-storage": "5.0.2",
-    "node-mailjet": "6.0.2",
-    "nodemailer": "6.9.1",
-    "nodemon": "2.0.22",
+    "node-mailjet": "6.0.3",
+    "nodemailer": "6.9.3",
+    "nodemon": "3.0.1",
     "passport": "0.6.0",
     "passport-jwt": "4.0.1",
     "reflect-metadata": "0.1.13",
     "rfdc": "1.3.0",
-    "rimraf": "4.4.1",
-    "rxjs": "7.8.0"
+    "rimraf": "5.0.1",
+    "rxjs": "7.8.1"
   },
   "devDependencies": {
-    "@nestjs/testing": "9.3.12",
-    "@types/compression": "1.7.2",
+    "@nestjs/cli": "10.1.8",
+    "@nestjs/schematics": "10.0.1",
+    "@nestjs/testing": "10.0.5",
+    "@swc/cli": "^0.1.62",
+    "@swc/core": "1.3.69",
+    "@swc/jest": "0.2.26",
+    "@types/compression": "^1.7.2",
     "@types/cookie-parser": "1.4.3",
     "@types/cron": "2.0.1",
     "@types/ejs": "3.1.2",
-    "@types/jest": "29.5.0",
-    "@types/lodash": "4.14.192",
+    "@types/express": "4.17.17",
+    "@types/jest": "29.5.3",
+    "@types/lodash": "4.14.195",
     "@types/multer": "1.4.7",
-    "@types/node": "18.15.10",
-    "@types/nodemailer": "6.4.7",
+    "@types/node": "20.4.2",
+    "@types/nodemailer": "6.4.8",
     "@types/passport": "1.0.12",
     "@types/supertest": "2.0.12",
-    "@typescript-eslint/eslint-plugin": "5.57.0",
-    "@typescript-eslint/parser": "5.57.0",
+    "@typescript-eslint/eslint-plugin": "6.0.0",
+    "@typescript-eslint/parser": "6.0.0",
     "coffeescript": "2.7.0",
-    "eslint": "8.36.0",
+    "eslint": "8.45.0",
     "eslint-config-prettier": "8.8.0",
     "find-file-up": "2.0.1",
     "grunt": "1.6.1",
@@ -126,17 +134,18 @@
     "grunt-contrib-watch": "1.1.0",
     "grunt-sync": "0.8.2",
     "husky": "8.0.3",
-    "jest": "29.5.0",
+    "jest": "29.6.1",
     "npm-watch": "0.11.0",
     "pm2": "5.3.0",
-    "prettier": "2.8.7",
+    "prettier": "2.8.8",
     "pretty-quick": "3.1.3",
     "supertest": "6.3.3",
-    "ts-jest": "29.0.5",
-    "ts-morph": "17.0.1",
+    "ts-jest": "29.1.1",
+    "ts-loader": "9.4.4",
+    "ts-morph": "19.0.0",
     "ts-node": "10.9.1",
-    "tsconfig-paths": "4.1.2",
-    "typescript": "5.0.2",
+    "tsconfig-paths": "4.2.0",
+    "typescript": "5.1.6",
     "yalc": "1.0.0-pre.53"
   },
   "overrides": {
