@@ -1,8 +1,8 @@
 import { NestFactory } from '@nestjs/core';
 import { NestExpressApplication } from '@nestjs/platform-express';
 import { exec } from 'child_process';
-import * as compression from 'compression';
-import * as cookieParser from 'cookie-parser';
+import compression = require('compression');
+import cookieParser = require('cookie-parser');
 import envConfig from './config.env';
 import { HttpExceptionLogFilter } from './core/common/filters/http-exception-log.filter';
 import { ServerModule } from './server/server.module';
